+++ src/Implicit/UserItem.cs
@@ -2,7 +2,7 @@
 {
     public struct UserItem
     {
-        public UserItem(string userId, string itemId, double confidence)
+        public UserItem(string userId, string itemId, float confidence)
         {
             this.UserId = userId;
             this.ItemId = itemId;
@@ -13,6 +13,6 @@
 
         public string ItemId { get; }
 
-        public double Confidence { get; }
+        public float Confidence { get; }
     }
 }
